Converted it into a maven project. - API and plugin is now split into parts.
