+++ API/src/main/java/com/github/robotnikthingy/weaponsapi/WeaponsAPIInterface.java
@@ -1,0 +1,7 @@
+package com.github.robotnikthingy.weaponsapi;
+
+/**
+ * Created by ysl3000
+ */
+public interface WeaponsAPIInterface {
+}
