+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/ActionSystem/ActionParser.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.ActionSystem;
+
+/**
+ * Created by Robotnik on 8/9/2017.
+ * Will be used to parse actions and their parameters from the config files
+ */
+public class ActionParser {
+}
