+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/ActionSystem/Actions/ActionManager.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.ActionSystem.Actions;
+
+/**
+ * Created by Robotnik on 8/9/2017.
+ * Will manage all config related actions used by the plugin. EX: FireProjectile, ParticleEffect, PlaySound
+ */
+public class ActionManager {
+}
