+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/ActionSystem/Events/EventManager.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.ActionSystem.Events;
+
+/**
+ * Created by Robotnik on 8/9/2017.
+ * Will manage all config related events used by the plugin EX: OnReload, OnHit, etc
+ */
+public class EventManager {
+}
