+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/ActionSystem/Requirements/RequirementManager.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.ActionSystem.Requirements;
+
+/**
+ * Created by Robotnik on 8/9/2017.
+ * Will manage the requirements as defined in the config files. EX: DistanceFromShooter, Requires_Ammo, Required_Items...
+ */
+public class RequirementManager {
+}
