+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/ActionSystem/TargetAction/TargetManager.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.ActionSystem.TargetAction;
+
+/**
+ * Created by Robotnik on 8/9/2017.
+ * Will manage the targets for actions as defined in the config files. EX: SelfActions, VictimActions, NonTargetableActions, etc
+ */
+public class TargetManager {
+}
