+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/Ammo/AmmoItem.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.Ammo;
+
+/**
+ * Created by Robotnik on 8/9/2017.
+ * This represents an ammo item and all its properties
+ */
+public class AmmoItem {
+}
