+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/Attachments/AttachmentItem.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.Attachments;
+
+/**
+ * Created by Robotnik on 8/9/2017.
+ * This class represents an attachment and its properties
+ */
+public class AttachmentItem {
+}
