+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/Explosion/CustomExplosion.java
@@ -1,0 +1,7 @@
+package com.github.robotnikthingy.weaponsapi.Explosion;
+
+/**
+ * Created by Robotnik on 8/4/2017.
+ */
+public class CustomExplosion {
+}
