+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/Gui/LoreManager.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.Gui;
+
+/**
+ * Created by Robotnik on 8/9/2017.
+ * Will handle managing and updating lore related to the plugin
+ */
+public class LoreManager {
+}
