+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/Projectile/WeaponProjectile.java
@@ -1,0 +1,7 @@
+package com.github.robotnikthingy.weaponsapi.Projectile;
+
+/**
+ * Created by Robotnik on 8/4/2017.
+ */
+public class WeaponProjectile {
+}
