+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/Weapons/WeaponItem.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.Weapons;
+
+/**
+ * Created by Robotnik on 8/4/2017.
+ * This represents a weapon and all its properties
+ */
+public class WeaponItem {
+}
