+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/Weapons/WeaponManager.java
@@ -1,0 +1,8 @@
+package com.github.robotnikthingy.weaponsapi.Weapons;
+
+/**
+ * Created by Robotnik on 8/9/2017.
+ * Will be used to keep track of and manage all weapons loaded by the plugin
+ */
+public class WeaponManager {
+}
