+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/WeaponsAPI.java
@@ -1,0 +1,13 @@
+package com.github.robotnikthingy.weaponsapi;
+
+import org.bukkit.plugin.java.JavaPlugin;
+
+/**
+ * Created by Robotnik on 8/4/2017.
+ */
+public class WeaponsAPI extends JavaPlugin {
+
+    public void onEnable() {
+
+    }
+}
