+++ Implementation/src/main/java/com/github/robotnikthingy/weaponsapi/WeaponsAPIPlugin.java
@@ -1,0 +1,12 @@
+package com.github.robotnikthingy.weaponsapi;
+
+import org.bukkit.plugin.java.JavaPlugin;
+
+/**
+ * Created by ysl3000
+ */
+public class WeaponsAPIPlugin extends JavaPlugin{
+
+
+
+}
