+++ Implementation/src/main/resources/ExampleAmmoItemConfig.yml
