+++ Implementation/src/main/resources/ExampleExplosionsConfig.yml
