+++ Implementation/src/main/resources/ExampleGunsConfig.yml
