+++ Implementation/src/main/resources/ExampleProjectileConfig.yml
