--- docs/Javadocs_will_go_here.txt
@@ -1,1 +1,0 @@
-Javadocs go here
