--- src/WeaponsApi/ActionSystem/ActionParser.java
@@ -1,8 +1,0 @@
-package WeaponsApi.ActionSystem;
-
-/**
- * Created by Robotnik on 8/9/2017.
- * Will be used to parse actions and their parameters from the config files
- */
-public class ActionParser {
-}
