--- src/WeaponsApi/ActionSystem/Actions/ActionManager.java
@@ -1,8 +1,0 @@
-package WeaponsApi.ActionSystem.Actions;
-
-/**
- * Created by Robotnik on 8/9/2017.
- * Will manage all config related actions used by the plugin. EX: FireProjectile, ParticleEffect, PlaySound
- */
-public class ActionManager {
-}
