--- src/WeaponsApi/ActionSystem/Events/EventManager.java
@@ -1,8 +1,0 @@
-package WeaponsApi.ActionSystem.Events;
-
-/**
- * Created by Robotnik on 8/9/2017.
- * Will manage all config related events used by the plugin EX: OnReload, OnHit, etc
- */
-public class EventManager {
-}
