--- src/WeaponsApi/ActionSystem/Requirements/RequirementManager.java
@@ -1,8 +1,0 @@
-package WeaponsApi.ActionSystem.Requirements;
-
-/**
- * Created by Robotnik on 8/9/2017.
- * Will manage the requirements as defined in the config files. EX: DistanceFromShooter, Requires_Ammo, Required_Items...
- */
-public class RequirementManager {
-}
