--- src/WeaponsApi/ActionSystem/TargetAction/TargetManager.java
@@ -1,8 +1,0 @@
-package WeaponsApi.ActionSystem.TargetAction;
-
-/**
- * Created by Robotnik on 8/9/2017.
- * Will manage the targets for actions as defined in the config files. EX: SelfActions, VictimActions, NonTargetableActions, etc
- */
-public class TargetManager {
-}
