--- src/WeaponsApi/Ammo/AmmoItem.java
@@ -1,8 +1,0 @@
-package WeaponsApi.Ammo;
-
-/**
- * Created by Robotnik on 8/9/2017.
- * This represents an ammo item and all its properties
- */
-public class AmmoItem {
-}
