--- src/WeaponsApi/Attachments/AttachmentItem.java
@@ -1,8 +1,0 @@
-package WeaponsApi.Attachments;
-
-/**
- * Created by Robotnik on 8/9/2017.
- * This class represents an attachment and its properties
- */
-public class AttachmentItem {
-}
