--- src/WeaponsApi/Explosion/CustomExplosion.java
@@ -1,7 +1,0 @@
-package WeaponsApi.Explosion;
-
-/**
- * Created by Robotnik on 8/4/2017.
- */
-public class CustomExplosion {
-}
