--- src/WeaponsApi/Gui/LoreManager.java
@@ -1,8 +1,0 @@
-package WeaponsApi.Gui;
-
-/**
- * Created by Robotnik on 8/9/2017.
- * Will handle managing and updating lore related to the plugin
- */
-public class LoreManager {
-}
