--- src/WeaponsApi/Projectile/WeaponProjectile.java
@@ -1,7 +1,0 @@
-package WeaponsApi.Projectile;
-
-/**
- * Created by Robotnik on 8/4/2017.
- */
-public class WeaponProjectile {
-}
