--- src/WeaponsApi/Weapons/WeaponItem.java
@@ -1,8 +1,0 @@
-package WeaponsApi.Weapons;
-
-/**
- * Created by Robotnik on 8/4/2017.
- * This represents a weapon and all its properties
- */
-public class WeaponItem {
-}
