--- src/WeaponsApi/Weapons/WeaponManager.java
@@ -1,8 +1,0 @@
-package WeaponsApi.Weapons;
-
-/**
- * Created by Robotnik on 8/9/2017.
- * Will be used to keep track of and manage all weapons loaded by the plugin
- */
-public class WeaponManager {
-}
