--- src/WeaponsApi/WeaponsAPI.java
@@ -1,13 +1,0 @@
-package WeaponsApi;
-
-import org.bukkit.plugin.java.JavaPlugin;
-
-/**
- * Created by Robotnik on 8/4/2017.
- */
-public class WeaponsAPI extends JavaPlugin {
-
-    public void onEnable() {
-
-    }
-}
